Parse flamdex Query Term's from JSON using longValue() instead of intValue()

+++ src/java/com/indeed/squall/iql2/execution/actions/Actions.java
@@ -117,7 +117,7 @@ public class Actions {
         final String field = json.get("field").textValue();
         final boolean isIntField = json.get("isIntField").booleanValue();
         if (isIntField) {
-            return Term.intTerm(field, json.get("intTerm").intValue());
+            return Term.intTerm(field, json.get("intTerm").longValue());
         } else {
             return Term.stringTerm(field, json.get("stringTerm").textValue());
         }
